Refactorisation EditProfileUi.cs : extraction de la barre d'outils BBCode en composant réutilisable `BbCodeToolbar`.
